Fix bug introduced by bs4 4.10+.
It appears that Beautiful Soup 4.10 is parsing HTML slightly
differently. As a result, our parsed data now includes an
extra element: The first element is now an empty string,
rather than the rank.

That is, w/ 4.9.x bs4 would return the following for the first
data row:

    ['1', 'gonz', ...]

But in 4.10.x it now returns

    ['', '1', 'gonz', ...]

The solution, for now, is to omit any blank text elements.